Move currentlyInstantiatingEntity, register from hydrate.
