+++ packages/orm/src/BaseEntity.ts
@@ -12,6 +12,8 @@ import {
   TaggedId,
 } from "./index";
 
+export let currentlyInstantiatingEntity: Entity | undefined;
+
 /**
  * The base class for all entities.
  *
@@ -19,27 +21,23 @@ import {
  */
 export abstract class BaseEntity<EM extends EntityManager, I extends IdType = IdType> implements Entity {
   readonly __orm!: EntityOrmField;
+
+  protected constructor(em: EM, metadata: any, defaultValues: object, optsOrId: any) {
+    // Only do em.register for em.create-d entities, otherwise defer to hydrate to em.register
+    if (typeof optsOrId === "string") {
+      this.__orm = new EntityOrmField(em, metadata, undefined);
+    } else {
+      this.__orm = new EntityOrmField(em, metadata, defaultValues);
+      em.register(this);
+    }
+    currentlyInstantiatingEntity = this;
+  }
+
   // This gives rules a way to access the fully typed object instead of their Reacted view.
   // And we make it public so that a function that takes Reacted<...> can accept a Loaded<...>
   // that sufficiently overlaps.
-  readonly fullNonReactiveAccess!: this;
-
-  protected constructor(em: EM, metadata: any, defaultValues: object, optsOrId: any) {
-    Object.defineProperty(this, "__orm", {
-      value: new EntityOrmField(em, metadata, defaultValues),
-      enumerable: false,
-    });
-    Object.defineProperty(this, "fullNonReactiveAccess", {
-      value: this,
-      enumerable: false,
-      writable: false,
-    });
-    // Ensure we have at least id set so the `EntityManager.register` works
-    if (typeof optsOrId === "string") {
-      this.__orm.data["id"] = optsOrId;
-      this.__orm.isNew = false;
-    }
-    em.register(metadata, this);
+  get fullNonReactiveAccess(): this {
+    return this;
   }
 
   /** @returns the entity's id, tagged/untagged based on your config, or a runtime error if it's new/unassigned. */
